PROBLEM: Don't know what string literals are used for ASSET operation. There is no protocol definition. SOLUTION: Add it.

+++ src/bios_proto.c
@@ -84,7 +84,7 @@ struct _bios_proto_t {
 
 //  Put a 1-byte number to the frame
 #define PUT_NUMBER1(host) { \
-    *(byte *) self->needle = (host); \
+    *(byte *) self->needle = (byte) (host); \
     self->needle++; \
 }
